Added GamePlay Screen

+++ core/src/main/java/io/github/AngryBird/BrownBlock.java
@@ -5,7 +5,7 @@ import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 
 public class BrownBlock {
     Texture img=new Texture("brownBlock.png");
-    public void draw(SpriteBatch batch, int x, int y, int sizeX, int sizeY) {
+    public void draw(SpriteBatch batch, float x, float y, float sizeX, float sizeY) {
         batch.draw(img,x,y,sizeX,sizeY);
     }
 }
