+++ core/src/main/java/io/github/AngryBird/Pig.java
@@ -9,7 +9,7 @@ public class Pig extends BasePig{
     public Pig(int x, int y) {
         super(x, y);
     }
-    public void draw(SpriteBatch batch, int x, int y, int sizeX, int sizeY) {
+    public void draw(SpriteBatch batch, float x, float y, float sizeX, float sizeY) {
         batch.draw(img,x,y,sizeX,sizeY);
     }
 }
