+++ core/src/main/java/io/github/AngryBird/YellowBird.java
@@ -8,7 +8,7 @@ public class YellowBird extends Bird {
     public YellowBird(int x, int y) {
         super(x, y);
     }
-    public void draw(SpriteBatch batch, int x, int y, int sizeX, int sizeY) {
+    public void draw(SpriteBatch batch, float x, float y, float sizeX, float sizeY) {
         batch.draw(img,x,y,sizeX,sizeY);
     }
 }
